requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements/base.txt requirements/base.in
#
annotated-types==0.7.0
    # via pydantic
certifi==2025.8.3
    # via
    #   clickhouse-connect
    #   requests
charset-normalizer==3.4.3
    # via requests
clickhouse-connect==0.8.18
    # via -r requirements/base.in
idna==3.10
    # via requests
loguru==0.7.3
    # via -r requirements/base.in
lz4==4.4.4
    # via clickhouse-connect
numpy==2.3.2
    # via
    #   -r requirements/base.in
    #   pandas
pandas==2.3.1
    # via -r requirements/base.in
psycopg2-binary==2.9.10
    # via -r requirements/base.in
pydantic==2.11.7
    # via -r requirements/base.in
pydantic-core==2.33.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via -r requirements/base.in
pytz==2025.2
    # via
    #   clickhouse-connect
    #   pandas
requests==2.32.5
    # via -r requirements/base.in
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.43
    # via -r requirements/base.in
typing-extensions==4.14.1
    # via
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   clickhouse-connect
    #   requests
zstandard==0.24.0
    # via clickhouse-connect
